refactor frontend integrations code out of settings and into its own manager class, backend paperless integration code to its own integrations file which contains everything it needs (schema, functions, endpoints)
