+++ server.js
@@ -26,7 +26,7 @@ const { sanitizeFileName } = require('./src/services/fileUpload/utils');
 const packageJson = require('./package.json');
 const { TOKENMASK } = require('./src/constants');
 const { integrationManager } = require('./integrations/integrationManager');
-const PaperlessEndpoints = require('./integrations/api/paperlessEndpoints');
+const PaperlessIntegration = require('./integrations/paperless');
 
 const app = express();
 const PORT = process.env.PORT || 3000;
@@ -2190,7 +2190,7 @@ app.post('/api/settings', (req, res) => {
 
 // --- INTEGRATION SYSTEM ---
 // Initialize and register integration routes
-PaperlessEndpoints.registerRoutes(app, getAppSettings);
+PaperlessIntegration.registerRoutes(app, getAppSettings);
 
 // Test notification endpoint
 app.post('/api/notification-test', async (req, res) => {
